refactor(spec): split recipe artifact modules
